Fix GH build.

+++ radio/src/datastructs_private.h
@@ -93,7 +93,7 @@ PACK(struct MixData {
   uint16_t carryTrim:1;
   uint16_t mixWarn:2;       // mixer warning
   uint16_t mltpx:2 ENUM(MixerMultiplex);
-  uint32_t delayPrec:1;
+  uint16_t delayPrec:1;
   uint16_t speedPrec:1;
   uint16_t flightModes:9 CUST(r_flightModes, w_flightModes);
   uint16_t spare:1 SKIP;
